Builder: abstract prefetch as a core module
Move prefetch code to an individual module, let this part pull
away from the core build logic, make workflow clearer.
