+++ src/bin/nydus-image/core/mod.rs
@@ -1,0 +1,5 @@
+// Copyright 2020 Ant Group. All rights reserved.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod prefetch;
